Weapon Projectiles added, Ion functions, pathfinding technique
Weapons now fire projectiles that take time to reach their target. Ion weapons added to the mix, as well as system ionization. There is a new pathfinding function that will later be implemented for a possible 2d map remodeling. Crew will use it to get to rooms they are ordered to. Hopefully I can upgrade the ship to be more than a long line.